radicle: identity caching
Add caching for the `Identity` COB.

A migration file is added for the sqlite COB cache, which inserts the
new `identities` table.

A new module `identity::cache` is added, similar to `patch::cache` and
`issue::cache`.

To achieve the same pattern as patches and issues, it was necessary to
introduce a new type `Identities` in the `identity` module, which
offers the same functionality as methods that existed on the
`Identity` type.

The caching is implemented using write-through semantics, again
similar to patches and issues.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>
X-Clacks-Overhead: GNU Terry Pratchett