cli: improve passphrase input experience
When an incorrect passphrase is entered, the characters are not
removed. Unfortunately, this is due to the upstream library we are
using[[0]].

To improve the experience, however, the `with_display_toggle_enabled`
can be used to reveal the user's input using the Ctrl+R hotkey. They
can then reveal the password, if able to, and fix the problematic
input.

Once the feature request in the upstream library is made, that can be
used instead.

[0]: https://github.com/mikaelmello/inquire/issues/108#issuecomment-2015745056

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>
X-Clacks-Overhead: GNU Terry Pratchett