node: Add user-agent to node announcement
Adds a user-agent string to the node announcement. This lets us keep
track of what software version everyone is running. Especially useful to
see what percentage of the network has upgraded to a new release.

For old nodes, this defaults to the string `/radicle/`.

The user agent format is roughly based on BIP 0014.