cli: use git config's `core.abbrev` for SHA-1 abbreviations
Attempt to determine how to abbreviate SHA-1 values in the CLI by checking the
`core.abbrev` value in Git configuration.

If this is not set then a default of 7 is used as a fallback. Note that a value
of 10 is a better recommendation, however, due to testing values, 7 was chosen
to minimise the changes made in this commit. It should be considered, in the
future, to use `auto` as a default, and then 10 as a final fallback value (or
pehaps 40 -- the max).

Co-authored-by: Fintan Halpenny <fintan.halpenny@gmail.com>