cli: clear by notification state
Add the ability to clear a set of `rad inbox` items by a notifications state.

This is achieved by introducing a field to the `Clear` operation to optionally
clear by a state. When `rad inbox clear` is called, it will check if the
`--by-state` option is used, if not then the behaviour is as before, otherwise
it will get all notification ids matching the given state (and `Mode`)
before removing all those notifications.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>
X-Clacks-Overhead: GNU Terry Pratchett