radicle: add canonical refs to identity
The main change of this patch is adding canonical reference rules to the identity
document. Since this is a core piece to the stack it does come with a lot of
consequential changes.

A large majority of changes come from the fact that changing the contents of the
identity document changes the resulting SHAs of the identity. This has a knock
on effect on RIDs in the tests, and SHAs of COBs since they store the identity
SHA alongside their data. Other changes in the CLI example tests include:

  - Having the `canonicalRefs` field included in the identity output
  - Removal of `threshold` field in the identity output
  - Some copy changes for canonical references when pushing

The change to the `RawDoc` and `Doc` types follows from the Canonical References
RIP. The `RawRules` and `Rules` types are added to each, respectively. Since the
`threshold` is not used anymore, this is dropped from `Doc` and `RawDoc` – in
the case of the version 2 document.

For compatibility, if the document is version 1 and the repository is a project,
the verified `Doc` will use the existing `threshold` to ensure that there is a
rule for the default branch – which consists of the delegates and the threshold.

A test is introduced to try to ensure that migration from the old document
format to the latest document format is compatible. It does this by hardcoding
the old document value and seeing that it can deserialize into the latest
`RawDoc` and `Doc` types. Note that this should future-proof any further
changes, since this test will fail if any non-backwards-compatible changes are
made.

By removing the threshold from the identity, the threshold for a project had to
be obtained from the canonical reference rules. This is done by using a helper,
`default_branch_rule`, for getting the threshold if it exists. For
`canonical_head` this means that it will fail it could not find the rule.
Within `radicle-fetch` it means that verification using the threshold only
occurs when the threshold is available. Similarly, in `rad id`, the threshold
for verification only applies for project repositories and is skipped otherwise.

Following from removing `threshold`, this meant that some tests that used `rad
id update --threshold` needed to be changed. To do this a new command `rad
canonical-ref` was added to manage canonical reference rules. Some code was
extracted from `commands/id.rs` into a common module that could be reused across
both `commands/id.rs` and `commands/canonical-ref.rs`. Finally, the
`--threshold` option was removed since it was no longer relevant.

The more functional part of the work for using the canonical reference rules
happens in `radicle-node/src/worker/fetch.rs` and
`radicle-remote-helper/src/push.rs`. In the former, the updated references are
iterated over to check if they match any of the given rules in an identity. If a
match is found, the canonical reference is calculated and set. In the latter,
when a reference is pushed, the canonical references are also checked for a
match and the canonical reference is similarly calculated and set. Note that
this code is very similar to original default branch code, while augmenting it
for the rule matching.

An example test, `git-push-canonical-tags`, is added to demonstrate how
canonical reference rules can be used for tags -- which is deemed as one of the
important use cases of this feature.

One change of note is that the `rad-id-threshold` test no longer made any sense
to test with how the CLI works -- see the FIXME comment for more details. It is
ignored for the time being.