radicle: generic signing and verification
Attempt to use the `Verifier` and `Signer`, from the `signature`
crate, in places where we sign and verify – effectively making the
sign and verify sections of our code generic to plug in other
implementors of these traits.

Sketching out where NodeId and Did should be used, grep for
TODO(finto) to see notes about these.